pyproject.toml
[project]
name = "migrate"
version = "0.1.0"
description = "AI-assisted GCP → Databricks/AWS migration framework with human-in-the-loop"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "rich>=13.7",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
    "sqlglot>=25.0",
    "google-cloud-bigquery>=3.25",
    "databricks-sdk>=0.30",
    "anthropic>=0.40",
]

[project.scripts]
migrate = "migrate.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/migrate"]

[tool.hatch.build.targets.wheel.force-include]
"src/migrate/core/convert/_templates" = "migrate/core/convert/_templates"
"src/migrate/web/templates" = "migrate/web/templates"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "ruff>=0.5",
]
